@@ -1,17 +1,20 @@
 package io.github.jhipster.sample.config;
 
 import io.github.jhipster.config.JHipsterConstants;
+import io.github.jhipster.config.h2.H2ConfigurationHelper;
 import io.github.jhipster.config.liquibase.AsyncSpringLiquibase;
 
 import liquibase.integration.spring.SpringLiquibase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.liquibase.LiquibaseProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.core.env.Environment;
+
 import org.springframework.core.task.TaskExecutor;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
@@ -45,30 +48,7 @@ public class DatabaseConfiguration {
     @Bean(initMethod = "start", destroyMethod = "stop")
     @Profile(JHipsterConstants.SPRING_PROFILE_DEVELOPMENT)
     public Object h2TCPServer() throws SQLException {
-        try {
-            // We don't want to include H2 when we are packaging for the "prod" profile and won't
-            // actually need it, so we have to load / invoke things at runtime through reflection.
-            ClassLoader loader = Thread.currentThread().getContextClassLoader();
-            Class<?> serverClass = Class.forName("org.h2.tools.Server", true, loader);
-            Method createServer = serverClass.getMethod("createTcpServer", String[].class);
-            return createServer.invoke(null, new Object[] { new String[] { "-tcp", "-tcpAllowOthers" } });
-
-        } catch (ClassNotFoundException | LinkageError  e) {
-            throw new RuntimeException("Failed to load and initialize org.h2.tools.Server", e);
-
-        } catch (SecurityException | NoSuchMethodException e) {
-            throw new RuntimeException("Failed to get method org.h2.tools.Server.createTcpServer()", e);
-
-        } catch (IllegalAccessException | IllegalArgumentException e) {
-            throw new RuntimeException("Failed to invoke org.h2.tools.Server.createTcpServer()", e);
-
-        } catch (InvocationTargetException e) {
-            Throwable t = e.getTargetException();
-            if (t instanceof SQLException) {
-                throw (SQLException) t;
-            }
-            throw new RuntimeException("Unchecked exception in org.h2.tools.Server.createTcpServer()", t);
-        }
+        return H2ConfigurationHelper.createServer();
     }
 
     @Bean
