@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import promiseMiddleware from 'redux-promise-middleware';
 import thunkMiddleware from 'redux-thunk';
-import reducer from '../reducers';
+import reducer from 'app/shared/reducers';
 import DevTools from './devtools';
 import errorMiddleware from './error-middleware';
 import notificationMiddleware from './notification-middleware';
@@ -16,22 +16,11 @@ const defaultMiddlewares = [
   loadingBarMiddleware(),
   loggerMiddleware
 ];
-const composedMiddlewares = middlewares => process.env.NODE_ENV === 'development' ?
-  compose(applyMiddleware(...defaultMiddlewares, ...middlewares), DevTools.instrument()) :
-  compose(applyMiddleware(...defaultMiddlewares, ...middlewares));
+const composedMiddlewares = middlewares =>
+  process.env.NODE_ENV === 'development'
+    ? compose(applyMiddleware(...defaultMiddlewares, ...middlewares), DevTools.instrument())
+    : compose(applyMiddleware(...defaultMiddlewares, ...middlewares));
 
-const initialize = (initialState = {}, middlewares = []) => {
-  const store = createStore(reducer, initialState, composedMiddlewares(middlewares));
-
-  // Enable Webpack hot module replacement for reducers
-  if (module.hot) {
-    // TODO : see if reducers can be moved to feature modules and still get HMR working
-    module.hot.accept('../reducers', () => {
-      const nextReducer = require('../reducers').default;
-      store.replaceReducer(nextReducer);
-    });
-  }
-  return store;
-};
+const initialize = (initialState = {}, middlewares = []) => createStore(reducer, initialState, composedMiddlewares(middlewares));
 
 export default initialize;
