@@ -1,88 +1,121 @@
-/* eslint-disable */ // TODO Fix when page is completed
 import * as React from 'react';
+import { Translate, translate } from 'react-jhipster';
 import { connect } from 'react-redux';
-import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
+import { AvForm, AvField } from 'availity-reactstrap-validation';
+import { Row, Col, Button } from 'reactstrap';
 
-import { getSession } from '../../../reducers/authentication';
-import { savePassword } from '../../../reducers/account';
+import { getSession } from 'app/shared/reducers/authentication';
+import PasswordStrengthBar from 'app/shared/layout/password/password-strength-bar';
+import { savePassword, reset } from './password.reducer';
 
 export interface IUserSettingsProps {
   account: any;
   getSession: Function;
+  savePassword: Function;
+  reset: Function;
 }
 
 export interface IUserSettingsState {
-  account: any;
-  firstPassword: string;
-  secondPassword: string;
+  password: string;
 }
 
 export class SettingsPage extends React.Component<IUserSettingsProps, IUserSettingsState> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      account: props.account,
-      firstPassword: null,
-      secondPassword: null
-    };
-  }
+  state: IUserSettingsState = {
+    password: ''
+  };
 
   componentDidMount() {
+    this.props.reset();
     this.props.getSession();
   }
 
-  componentWillReceiveProps(nextProps) {
-    this.setState({
-      account: nextProps.account
-    });
+  componentWillUnmount() {
+    this.props.reset();
   }
 
-  setFirstAccountPassword = event => {
-    this.setState({
-      firstPassword: event.target.value
-    });
-  }
+  handleValidSubmit = (event, values) => {
+    this.props.savePassword(values.currentPassword, values.newPassword);
+  };
 
-  setSecondAccountPassword = event => {
-    this.setState({
-      secondPassword: event.target.value
-    });
-  }
-
-  savePassword = event => {
-    if (this.state.firstPassword === this.state.secondPassword) {
-      savePassword(this.state.firstPassword);
-    }
-    event.preventDefault();
-  }
+  updatePassword = event => {
+    this.setState({ password: event.target.value });
+  };
 
   render() {
-    const { account } = this.state;
+    const { account } = this.props;
+
     return (
-        <div>
-          <h2>Password for [{account.login}]</h2>
-          <Form>
-            <FormGroup>
-              <Label>New password</Label>
-              <Input type="password" className="form-control" id="firstName" name="firstName" placeholder="New password"
-                     onChange={this.setFirstAccountPassword} />
-            </FormGroup>
-            <FormGroup>
-              <Label>New password confirmation</Label>
-              <Input type="password" className="form-control" id="lastName" name="lastName" placeholder="Confirm the new password" onChange={this.setSecondAccountPassword}/>
-            </FormGroup>
-             <Button type="submit" color="success" onClick={this.savePassword}>Save</Button>
-          </Form>
-        </div>
+      <div>
+        <Row className="justify-content-center">
+          <Col md="8">
+            <h2>
+              <Translate contentKey="password.title" interpolate={{ username: account.login }}>
+                Password for {account.login}
+              </Translate>
+            </h2>
+            <AvForm onValidSubmit={this.handleValidSubmit}>
+              <AvField
+                name="currentPassword"
+                label={translate('global.form.currentpassword')}
+                placeholder={translate('global.form.currentpassword.placeholder')}
+                type="password"
+                validate={{
+                  required: { value: true, errorMessage: translate('global.messages.validate.newpassword.required') }
+                }}
+              />
+              <AvField
+                name="newPassword"
+                label={translate('global.form.newpassword')}
+                placeholder={translate('global.form.newpassword.placeholder')}
+                type="password"
+                validate={{
+                  required: { value: true, errorMessage: translate('global.messages.validate.newpassword.required') },
+                  minLength: { value: 4, errorMessage: translate('global.messages.validate.newpassword.minlength') },
+                  maxLength: { value: 50, errorMessage: translate('global.messages.validate.newpassword.maxlength') }
+                }}
+                onChange={this.updatePassword}
+              />
+              <PasswordStrengthBar password={this.state.password} />
+              <AvField
+                name="confirmPassword"
+                label={translate('global.form.confirmpassword')}
+                placeholder={translate('global.form.confirmpassword.placeholder')}
+                type="password"
+                validate={{
+                  required: {
+                    value: true,
+                    errorMessage: translate('global.messages.validate.confirmpassword.required')
+                  },
+                  minLength: {
+                    value: 4,
+                    errorMessage: translate('global.messages.validate.confirmpassword.minlength')
+                  },
+                  maxLength: {
+                    value: 50,
+                    errorMessage: translate('global.messages.validate.confirmpassword.maxlength')
+                  },
+                  match: {
+                    value: 'newPassword',
+                    errorMessage: translate('global.messages.error.dontmatch')
+                  }
+                }}
+              />
+              <Button color="success" type="submit">
+                <Translate contentKey="password.form.button">Save</Translate>
+              </Button>
+            </AvForm>
+          </Col>
+        </Row>
+      </div>
     );
   }
 }
 
-const mapStateToProps = storeState => ({
-  account: storeState.authentication.account,
-  isAuthenticated: storeState.authentication.isAuthenticated
+const mapStateToProps = ({ authentication, password }) => ({
+  account: authentication.account,
+  isAuthenticated: authentication.isAuthenticated
 });
 
-const mapDispatchToProps = { getSession, savePassword };
+const mapDispatchToProps = { getSession, savePassword, reset };
 
 export default connect(mapStateToProps, mapDispatchToProps)(SettingsPage);
