@@ -1,11 +1,11 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Translate } from 'react-jhipster';
-import { Table } from 'reactstrap';
+import { Table, Badge } from 'reactstrap';
 import { FaEye, FaRefresh } from 'react-icons/lib/fa';
 
+import { systemHealth } from '../administration.reducer';
 import HealthModal from './health-modal';
-import { systemHealth } from '../../../reducers/administration';
 
 export interface IHealthPageProps {
   isFetching?: boolean;
@@ -14,15 +14,16 @@ export interface IHealthPageProps {
   systemHealthInfo: any;
 }
 
-export class HealthPage extends React.Component<IHealthPageProps, any> {
+export interface IHealthPageState {
+  healthObject: any;
+  showModal: boolean;
+}
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      healthObject: {},
-      showModal: false
-    };
-  }
+export class HealthPage extends React.Component<IHealthPageProps, IHealthPageState> {
+  state: IHealthPageState = {
+    healthObject: {},
+    showModal: false
+  };
 
   componentDidMount() {
     this.props.systemHealth();
@@ -32,7 +33,7 @@ export class HealthPage extends React.Component<IHealthPageProps, any> {
     if (!this.props.isFetching) {
       this.props.systemHealth();
     }
-  }
+  };
 
   getSystemHealthInfo = (name, healthObject) => () => {
     this.setState({
@@ -42,59 +43,73 @@ export class HealthPage extends React.Component<IHealthPageProps, any> {
         name
       }
     });
-  }
+  };
 
   handleClose = () => {
     this.setState({
-        showModal: false
+      showModal: false
     });
-  }
+  };
 
   renderModal = () => {
     const { healthObject } = this.state;
-    return (
-      <HealthModal healthObject={healthObject} handleClose={this.handleClose} showModal={this.state.showModal} />
-    );
-  }
+    return <HealthModal healthObject={healthObject} handleClose={this.handleClose} showModal={this.state.showModal} />;
+  };
 
   render() {
     const { health, isFetching } = this.props;
-    const data = health || {};
+    const data = (health || {}).details || {};
     return (
       <div>
-          <h2>Health Checks</h2>
-          <p>
-            <button type="button" onClick={this.getSystemHealth} className={isFetching ? 'btn btn-danger' : 'btn btn-primary'} disabled={isFetching}>
-              <FaRefresh />&nbsp;
-              <Translate component="span" contentKey="health.refresh.button" />
-            </button>
-          </p>
-          <div className="row">
-            <div className="col-12">
+        <h2>Health Checks</h2>
+        <p>
+          <button
+            type="button"
+            onClick={this.getSystemHealth}
+            className={isFetching ? 'btn btn-danger' : 'btn btn-primary'}
+            disabled={isFetching}
+          >
+            <FaRefresh />&nbsp;
+            <Translate component="span" contentKey="health.refresh.button">
+              Refresh
+            </Translate>
+          </button>
+        </p>
+        <div className="row">
+          <div className="col-12">
             <Table bordered>
-               <thead>
-                 <tr>
-                   <th>Service Name</th>
-                   <th>Status</th>
-                   <th>Details</th>
-                 </tr>
-               </thead>
-               <tbody>
-              {Object.keys(data).map((configPropKey, configPropIndex) =>
-                (configPropKey !== 'status') ?
-                  <tr key={configPropIndex}>
-                    <td>{configPropKey}</td>
-                    <td><button type="button" className={data[configPropKey].status !== 'UP' ? 'btn btn-danger' : 'btn btn-success'}>{data[configPropKey].status}</button></td>
-                    <td><a onClick={this.getSystemHealthInfo(configPropKey, data[configPropKey])}><FaEye /></a></td>
-                  </tr>
-                : null
+              <thead>
+                <tr>
+                  <th>Service Name</th>
+                  <th>Status</th>
+                  <th>Details</th>
+                </tr>
+              </thead>
+              <tbody>
+                {Object.keys(data).map(
+                  (configPropKey, configPropIndex) =>
+                    configPropKey !== 'status' ? (
+                      <tr key={configPropIndex}>
+                        <td>{configPropKey}</td>
+                        <td>
+                          <Badge color={data[configPropKey].status !== 'UP' ? 'danger' : 'success'}>{data[configPropKey].status}</Badge>
+                        </td>
+                        <td>
+                          {data[configPropKey].details ? (
+                            <a onClick={this.getSystemHealthInfo(configPropKey, data[configPropKey])}>
+                              <FaEye />
+                            </a>
+                          ) : null}
+                        </td>
+                      </tr>
+                    ) : null
                 )}
               </tbody>
-             </Table>
-            </div>
+            </Table>
           </div>
-          {this.renderModal()}
         </div>
+        {this.renderModal()}
+      </div>
     );
   }
 }
