@@ -4,22 +4,25 @@ import { Link } from 'react-router-dom';
 import { Button } from 'reactstrap';
 import {
   Translate,
-  ICrudGetAction,
+  ICrudGetAllAction,
+  ICrudPutAction,
   TextFormat,
   JhiPagination,
   getPaginationItemsNumber,
   getSortState,
-  IPaginationState
+  IPaginationBaseState
 } from 'react-jhipster';
 import { FaPlus, FaEye, FaPencil, FaSort, FaTrash } from 'react-icons/lib/fa';
 
-import { getUsers } from '../../../reducers/user-management';
-import { APP_DATE_FORMAT } from '../../../config/constants';
-import { ITEMS_PER_PAGE } from '../../../shared/util/pagination.constants';
+import { APP_DATE_FORMAT } from 'app/config/constants';
+import { IUser } from 'app/shared/model/user.model';
+import { ITEMS_PER_PAGE } from 'app/shared/util/pagination.constants';
+import { getUsers, updateUser } from './user-management.reducer';
 
 export interface IUserManagementProps {
-  getUsers: ICrudGetAction;
-  users: any[];
+  getUsers: ICrudGetAllAction<IUser>;
+  updateUser: ICrudPutAction<IUser>;
+  users: IUser[];
   account: any;
   match: any;
   totalItems: 0;
@@ -27,25 +30,24 @@ export interface IUserManagementProps {
   location: any;
 }
 
-export class UserManagement extends React.Component<IUserManagementProps, IPaginationState> {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      ...getSortState(props.location, ITEMS_PER_PAGE)
-    };
-  }
+export class UserManagement extends React.Component<IUserManagementProps, IPaginationBaseState> {
+  state: IPaginationBaseState = {
+    ...getSortState(this.props.location, ITEMS_PER_PAGE)
+  };
 
   componentDidMount() {
     this.getUsers();
   }
 
   sort = prop => () => {
-    this.setState({
-      order: this.state.order === 'asc' ? 'desc' : 'asc',
-      sort: prop
-    }, () => this.sortUsers());
-  }
+    this.setState(
+      {
+        order: this.state.order === 'asc' ? 'desc' : 'asc',
+        sort: prop
+      },
+      () => this.sortUsers()
+    );
+  };
 
   sortUsers() {
     this.getUsers();
@@ -57,7 +59,14 @@ export class UserManagement extends React.Component<IUserManagementProps, IPagin
   getUsers = () => {
     const { activePage, itemsPerPage, sort, order } = this.state;
     this.props.getUsers(activePage - 1, itemsPerPage, `${sort},${order}`);
-  }
+  };
+
+  toggleActive = user => () => {
+    this.props.updateUser({
+      ...user,
+      activated: !user.activated
+    });
+  };
 
   render() {
     const { users, account, match, totalItems } = this.props;
@@ -66,91 +75,116 @@ export class UserManagement extends React.Component<IUserManagementProps, IPagin
         <h2>
           <Translate contentKey="userManagement.home.title">Users</Translate>
           <Link to={`${match.url}/new`} className="btn btn-primary float-right jh-create-entity">
-            <FaPlus /> <Translate contentKey="userManagement.home.createLabel" />
+            <FaPlus /> <Translate contentKey="userManagement.home.createLabel">Create a new user</Translate>
           </Link>
         </h2>
         <div className="table-responsive">
           <table className="table table-striped">
             <thead>
               <tr>
-                <th onClick={this.sort('id')}><Translate contentKey="global.field.id">ID</Translate><FaSort/></th>
-                <th onClick={this.sort('login')}><Translate contentKey="userManagement.login">Login</Translate><FaSort/></th>
-                <th onClick={this.sort('email')}><Translate contentKey="userManagement.email">Email</Translate><FaSort/></th>
-                <th />
-                <th onClick={this.sort('langKey')}><Translate contentKey="userManagement.langKey">Lang Key</Translate><FaSort/></th>
-                <th><Translate contentKey="userManagement.profiles">Profiles</Translate></th>
-                <th onClick={this.sort('createdDate')}><Translate contentKey="userManagement.createdDate">Created Date</Translate><FaSort/></th>
-                <th onClick={this.sort('lastModifiedBy')}>
-                  <Translate contentKey="userManagement.lastModifiedBy">Last Modified By</Translate><FaSort/>
+                <th className="hand" onClick={this.sort('id')}>
+                  <Translate contentKey="global.field.id">ID</Translate>
+                  <FaSort />
                 </th>
-                <th onClick={this.sort('lastModifiedDate')}>
-                  <Translate contentKey="userManagement.lastModifiedDate">Last Modified Date</Translate><FaSort/>
+                <th className="hand" onClick={this.sort('login')}>
+                  <Translate contentKey="userManagement.login">Login</Translate>
+                  <FaSort />
+                </th>
+                <th className="hand" onClick={this.sort('email')}>
+                  <Translate contentKey="userManagement.email">Email</Translate>
+                  <FaSort />
+                </th>
+                <th />
+                <th className="hand" onClick={this.sort('langKey')}>
+                  <Translate contentKey="userManagement.langKey">Lang Key</Translate>
+                  <FaSort />
+                </th>
+                <th>
+                  <Translate contentKey="userManagement.profiles">Profiles</Translate>
+                </th>
+                <th className="hand" onClick={this.sort('createdDate')}>
+                  <Translate contentKey="userManagement.createdDate">Created Date</Translate>
+                  <FaSort />
+                </th>
+                <th className="hand" onClick={this.sort('lastModifiedBy')}>
+                  <Translate contentKey="userManagement.lastModifiedBy">Last Modified By</Translate>
+                  <FaSort />
+                </th>
+                <th className="hand" onClick={this.sort('lastModifiedDate')}>
+                  <Translate contentKey="userManagement.lastModifiedDate">Last Modified Date</Translate>
+                  <FaSort />
                 </th>
                 <th />
               </tr>
             </thead>
             <tbody>
-              {
-              users.map((user, i) => (
+              {users.map((user, i) => (
                 <tr key={`user-${i}`}>
                   <td>
-                    <Button
-                      tag={Link} to={`${match.url}/${user.login}`}
-                      color="link" size="sm"
-                    >
+                    <Button tag={Link} to={`${match.url}/${user.login}`} color="link" size="sm">
                       {user.id}
                     </Button>
                   </td>
                   <td>{user.login}</td>
                   <td>{user.email}</td>
                   <td>
-                    {
-                      user.activated ? (
-                        <span className="badge badge-success" style={{ cursor: 'pointer' }}>Activated</span>
-                      ) : (
-                        <span className="badge badge-danger" style={{ cursor: 'pointer' }}>Deactivated</span>
-                      )
-                    }
+                    {user.activated ? (
+                      <Button color="success" onClick={this.toggleActive(user)}>
+                        Activated
+                      </Button>
+                    ) : (
+                      <Button color="danger" onClick={this.toggleActive(user)}>
+                        Deactivated
+                      </Button>
+                    )}
                   </td>
                   <td>{user.langKey}</td>
                   <td>
-                    {
-                      user.authorities ? (
-                      user.authorities.map((authority, j) => (
-                        <div key={`user-auth-${i}-${j}`}>
-                          <span className="badge badge-info">{authority}</span>
-                        </div>
-                      ))) : null
-                    }
+                    {user.authorities
+                      ? user.authorities.map((authority, j) => (
+                          <div key={`user-auth-${i}-${j}`}>
+                            <span className="badge badge-info">{authority}</span>
+                          </div>
+                        ))
+                      : null}
                   </td>
-                  <td><TextFormat value={user.createdDate} type="date" format={APP_DATE_FORMAT} blankOnInvalid /></td>
+                  <td>
+                    <TextFormat value={user.createdDate} type="date" format={APP_DATE_FORMAT} blankOnInvalid />
+                  </td>
                   <td>{user.lastModifiedBy}</td>
-                  <td><TextFormat value={user.lastModifiedDate} type="date" format={APP_DATE_FORMAT} blankOnInvalid /></td>
+                  <td>
+                    <TextFormat value={user.lastModifiedDate} type="date" format={APP_DATE_FORMAT} blankOnInvalid />
+                  </td>
                   <td className="text-right">
                     <div className="btn-group flex-btn-group-container">
-                      <Button
-                        tag={Link} to={`${match.url}/${user.login}`}
-                        color="info" size="sm"
-                      >
-                        <FaEye/> <span className="d-none d-md-inline" ><Translate contentKey="entity.action.view" /></span>
+                      <Button tag={Link} to={`${match.url}/${user.login}`} color="info" size="sm">
+                        <FaEye />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.view">View</Translate>
+                        </span>
+                      </Button>
+                      <Button tag={Link} to={`${match.url}/${user.login}/edit`} color="primary" size="sm">
+                        <FaPencil />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.edit">Edit</Translate>
+                        </span>
                       </Button>
                       <Button
-                        tag={Link} to={`${match.url}/${user.login}/edit`}
-                        color="primary" size="sm"
+                        tag={Link}
+                        to={`${match.url}/${user.login}/delete`}
+                        color="danger"
+                        size="sm"
+                        disabled={account.login === user.login}
                       >
-                        <FaPencil/> <span className="d-none d-md-inline"><Translate contentKey="entity.action.edit" /></span>
-                      </Button>
-                      <Button
-                        tag={Link} to={`${match.url}/${user.login}/delete`}
-                        color="danger" size="sm" disabled={account.login === user.login}
-                      >
-                        <FaTrash/> <span className="d-none d-md-inline"><Translate contentKey="entity.action.delete" /></span>
+                        <FaTrash />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.delete">Delete</Translate>
+                        </span>
                       </Button>
                     </div>
                   </td>
                 </tr>
-              ))
-            }
+              ))}
             </tbody>
           </table>
         </div>
@@ -173,6 +207,6 @@ const mapStateToProps = storeState => ({
   account: storeState.authentication.account
 });
 
-const mapDispatchToProps = { getUsers };
+const mapDispatchToProps = { getUsers, updateUser };
 
 export default connect(mapStateToProps, mapDispatchToProps)(UserManagement);
