@@ -5,7 +5,7 @@ import { Link } from 'react-router-dom';
 import { Translate } from 'react-jhipster';
 import { connect } from 'react-redux';
 
-import { getSession } from '../../reducers/authentication';
+import { getSession } from 'app/shared/reducers/authentication';
 
 export interface IHomeProp {
   account: any;
@@ -17,12 +17,9 @@ export interface IHomeState {
 }
 
 export class Home extends React.Component<IHomeProp, IHomeState> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      currentUser: props.account
-    };
-  }
+  state: IHomeState = {
+    currentUser: this.props.account
+  };
 
   componentWillMount() {
     this.props.getSession();
@@ -36,36 +33,45 @@ export class Home extends React.Component<IHomeProp, IHomeState> {
 
   render() {
     const { currentUser } = this.state;
-
     return (
       <div className="row">
         <div className="col-md-9">
-          <h2><Translate contentKey="home.title">Welcome, Java Hipster!</Translate></h2>
-          <p className="lead"><Translate contentKey="home.subtitle">This is your homepage</Translate></p>
-          {
-            (currentUser && currentUser.login) ? (
-              <div>
-                <div className="alert alert-success">
-                  <Translate contentKey="home.logged.message" interpolate={{ username: currentUser.login }}>You are logged in as user </Translate>
-                </div>
+          <h2>
+            <Translate contentKey="home.title">Welcome, Java Hipster!</Translate>
+          </h2>
+          <p className="lead">
+            <Translate contentKey="home.subtitle">This is your homepage</Translate>
+          </p>
+          {currentUser && currentUser.login ? (
+            <div>
+              <div className="alert alert-success">
+                <Translate contentKey="home.logged.message" interpolate={{ username: currentUser.login }}>
+                  You are logged in as user {currentUser.login}.
+                </Translate>
               </div>
-            ) : (
-              <div>
-                <div className="alert alert-warning">
-                  <Translate contentKey="global.messages.info.authenticated.prefix">If you want to </Translate>
-                  <Link to="/login" className="alert-link"><Translate contentKey="global.messages.info.authenticated.link">sign in</Translate></Link>
-                  <Translate contentKey="global.messages.info.authenticated.suffix">, you can try the default accounts:
+            </div>
+          ) : (
+            <div>
+              <div className="alert alert-warning">
+                <Translate contentKey="global.messages.info.authenticated.prefix">If you want to </Translate>
+                <Link to="/login" className="alert-link">
+                  <Translate contentKey="global.messages.info.authenticated.link">sign in</Translate>
+                </Link>
+                <Translate contentKey="global.messages.info.authenticated.suffix">
+                  , you can try the default accounts:
                   <br />- Administrator (login=&quot;admin&quot; and password=&quot;admin&quot;)
-                  <br />- User (login=&quot;user&quot; and password=&quot;user&quot;).</Translate>
-                </div>
-
-                <div className="alert alert-warning">
-                  <Translate contentKey="global.messages.info.register.noaccount">You do not have an account yet?</Translate>&nbsp;
-                  <a className="alert-link"><Translate contentKey="global.messages.info.register.link">Register a new account</Translate></a>
-                </div>
+                  <br />- User (login=&quot;user&quot; and password=&quot;user&quot;).
+                </Translate>
               </div>
-            )
-          }
+
+              <div className="alert alert-warning">
+                <Translate contentKey="global.messages.info.register.noaccount">You do not have an account yet?</Translate>&nbsp;
+                <Link to="/register" className="alert-link">
+                  <Translate contentKey="global.messages.info.register.link">Register a new account</Translate>
+                </Link>
+              </div>
+            </div>
+          )}
           <p>
             <Translate contentKey="home.question">If you have any question on JHipster:</Translate>
           </p>
@@ -99,12 +105,14 @@ export class Home extends React.Component<IHomeProp, IHomeState> {
           </ul>
 
           <p>
-            <Translate contentKey="home.like">If you like JHipster, do not forget to give us a star on</Translate>
-            <a href="https://github.com/jhipster/generator-jhipster" target="_blank" rel="noopener noreferrer">Github</a>!
+            <Translate contentKey="home.like">If you like JHipster, do not forget to give us a star on</Translate>{' '}
+            <a href="https://github.com/jhipster/generator-jhipster" target="_blank" rel="noopener noreferrer">
+              Github
+            </a>!
           </p>
         </div>
         <div className="col-md-3 pad">
-          <span className="hipster img-fluid rounded" />
+          <span className="hipster rounded" />
         </div>
       </div>
     );
