@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { Translate, translate } from 'react-jhipster';
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Label } from 'reactstrap';
+import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Label, Alert } from 'reactstrap';
 import { AvForm, AvField, AvGroup, AvInput } from 'availity-reactstrap-validation';
+import { Link } from 'react-router-dom';
 
 export interface ILoginModalProps {
   showModal: boolean;
@@ -10,59 +11,78 @@ export interface ILoginModalProps {
   handleClose: Function;
 }
 
-class LoginModal extends React.Component<ILoginModalProps, {}> {
-
-  constructor(props, context) {
-    super(props, context);
-    this.state = {};
-  }
-
+class LoginModal extends React.Component<ILoginModalProps> {
   handleSubmit = (event, errors, { username, password, rememberMe }) => {
     const { handleLogin } = this.props;
     handleLogin(username, password, rememberMe);
-  }
+  };
 
   render() {
     const { loginError, handleClose } = this.props;
 
     return (
-      <Modal isOpen={this.props.showModal} toggle={handleClose} backdrop="static" id="login-page">
+      <Modal isOpen={this.props.showModal} toggle={handleClose} backdrop="static" id="login-page" autoFocus={false}>
         <AvForm onSubmit={this.handleSubmit}>
-          <ModalHeader toggle={handleClose} id="login-title"><Translate contentKey="login.title" /></ModalHeader>
+          <ModalHeader toggle={handleClose} id="login-title">
+            <Translate contentKey="login.title">Sign in</Translate>
+          </ModalHeader>
           <ModalBody>
             <div className="row">
               <div className="col-md-12">
-                { loginError ?
+                {loginError ? (
                   <div className="alert alert-danger">
-                    <Translate contentKey="login.messages.error.authentication" />
+                    <Translate contentKey="login.messages.error.authentication">
+                      <strong>Failed to sign in!</strong> Please check your credentials and try again.
+                    </Translate>
                   </div>
-                  : null
-                }
+                ) : null}
               </div>
               <div className="col-md-12">
                 <AvField
                   name="username"
-                  label={<Translate contentKey="global.form.username" />}
+                  label={translate('global.form.username')}
                   placeholder={translate('global.form.username.placeholder')}
-                  required errorMessage="Username cannot be empty!"
+                  required
+                  errorMessage="Username cannot be empty!"
+                  autoFocus
                 />
                 <AvField
-                  name="password" type="password"
-                  label={<Translate contentKey="login.form.password" />}
+                  name="password"
+                  type="password"
+                  label={translate('login.form.password')}
                   placeholder={translate('login.form.password.placeholder')}
-                  required errorMessage="Password cannot be empty!"
+                  required
+                  errorMessage="Password cannot be empty!"
                 />
-                <AvGroup>
-                  <Label check inline for="rememberMe" className="form-check-label">
-                    <AvInput type="checkbox" name="rememberMe" /> <Translate contentKey="login.form.rememberme" />
+                <AvGroup check inline>
+                  <Label className="form-check-label">
+                    <AvInput type="checkbox" name="rememberMe" /> <Translate contentKey="login.form.rememberme">Remember me</Translate>
                   </Label>
                 </AvGroup>
               </div>
             </div>
+            <div className="mt-1">&nbsp;</div>
+            <Alert color="warning">
+              <Link to="/reset/request">
+                <Translate contentKey="login.password.forgot">Did you forget your password?</Translate>
+              </Link>
+            </Alert>
+            <Alert color="warning">
+              <span>
+                <Translate contentKey="global.messages.info.register.noaccount">You don't have an account yet?</Translate>
+              </span>{' '}
+              <Link to="/register">
+                <Translate contentKey="global.messages.info.register.link">Register a new account</Translate>
+              </Link>
+            </Alert>
           </ModalBody>
           <ModalFooter>
-            <Button color="secondary" onClick={handleClose} tabIndex="1"><Translate contentKey="entity.action.cancel" /></Button>{' '}
-            <Button color="primary" type="submit"><Translate contentKey="login.form.button" /></Button>
+            <Button color="secondary" onClick={handleClose} tabIndex="1">
+              <Translate contentKey="entity.action.cancel">Cancel</Translate>
+            </Button>{' '}
+            <Button color="primary" type="submit">
+              <Translate contentKey="login.form.button">Sign in</Translate>
+            </Button>
           </ModalFooter>
         </AvForm>
       </Modal>
