@@ -2,14 +2,13 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
-import { logout } from '../../reducers/authentication';
+import { logout } from 'app/shared/reducers/authentication';
 
 export interface ILogoutProps {
   logout: Function;
 }
 
-export class Logout extends React.Component<ILogoutProps, undefined> {
-
+export class Logout extends React.Component<ILogoutProps> {
   componentWillMount() {
     this.props.logout();
   }
@@ -17,10 +16,12 @@ export class Logout extends React.Component<ILogoutProps, undefined> {
   render() {
     return (
       <div className="p-5">
-        <h4>Logged out succesfully!</h4>
-        <Redirect to={{
-          pathname: '/'
-        }} />
+        <h4>Logged out successfully!</h4>
+        <Redirect
+          to={{
+            pathname: '/'
+          }}
+        />
       </div>
     );
   }
