@@ -6,7 +6,9 @@ import { Translate } from 'react-jhipster';
 const Footer = props => (
   <div className="footer page-content">
     <div className="row">
-      <p className="col-md-12"><Translate contentKey="footer">Your footer</Translate></p>
+      <p className="col-md-12">
+        <Translate contentKey="footer">Your footer</Translate>
+      </p>
     </div>
   </div>
 );
