@@ -1,6 +1,5 @@
 import * as moment from 'moment';
 
-import { APP_LOCAL_DATETIME_FORMAT } from '../../config/constants';
+import { APP_LOCAL_DATETIME_FORMAT } from 'app/config/constants';
 
-export const convertDateTimeFromServer = date =>
-  date ? moment(date).format(APP_LOCAL_DATETIME_FORMAT) : null;
+export const convertDateTimeFromServer = date => (date ? moment(date).format(APP_LOCAL_DATETIME_FORMAT) : null);
