@@ -18,7 +18,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpSession;
 import java.time.Instant;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -76,85 +75,6 @@ public class CustomAuditEventRepositoryIntTest {
     }
 
     @Test
-    public void testFindAfter() {
-        persistenceAuditEventRepository.save(testUserEvent);
-        persistenceAuditEventRepository.save(testOldUserEvent);
-
-        List<AuditEvent> events =
-            customAuditEventRepository.find(Date.from(testUserEvent.getAuditEventDate().minusSeconds(3600)));
-        assertThat(events).hasSize(1);
-        AuditEvent event = events.get(0);
-        assertThat(event.getPrincipal()).isEqualTo(testUserEvent.getPrincipal());
-        assertThat(event.getType()).isEqualTo(testUserEvent.getAuditEventType());
-        assertThat(event.getData()).containsKey("test-key");
-        assertThat(event.getData().get("test-key").toString()).isEqualTo("test-value");
-        assertThat(event.getTimestamp()).isEqualTo(Date.from(testUserEvent.getAuditEventDate()));
-    }
-
-    @Test
-    public void testFindByPrincipal() {
-        persistenceAuditEventRepository.save(testUserEvent);
-        persistenceAuditEventRepository.save(testOldUserEvent);
-        persistenceAuditEventRepository.save(testOtherUserEvent);
-
-        List<AuditEvent> events = customAuditEventRepository
-            .find("test-user", Date.from(testUserEvent.getAuditEventDate().minusSeconds(3600)));
-        assertThat(events).hasSize(1);
-        AuditEvent event = events.get(0);
-        assertThat(event.getPrincipal()).isEqualTo(testUserEvent.getPrincipal());
-        assertThat(event.getType()).isEqualTo(testUserEvent.getAuditEventType());
-        assertThat(event.getData()).containsKey("test-key");
-        assertThat(event.getData().get("test-key").toString()).isEqualTo("test-value");
-        assertThat(event.getTimestamp()).isEqualTo(Date.from(testUserEvent.getAuditEventDate()));
-    }
-
-    @Test
-    public void testFindByPrincipalNotNullAndAfterIsNull() {
-        persistenceAuditEventRepository.save(testUserEvent);
-        persistenceAuditEventRepository.save(testOtherUserEvent);
-
-        List<AuditEvent> events = customAuditEventRepository.find("test-user", null);
-        assertThat(events).hasSize(1);
-        assertThat(events.get(0).getPrincipal()).isEqualTo("test-user");
-    }
-
-    @Test
-    public void testFindByPrincipalIsNullAndAfterIsNull() {
-        persistenceAuditEventRepository.save(testUserEvent);
-        persistenceAuditEventRepository.save(testOtherUserEvent);
-
-        List<AuditEvent> events = customAuditEventRepository.find(null, null);
-        assertThat(events).hasSize(2);
-        assertThat(events).extracting("principal")
-            .containsExactlyInAnyOrder("test-user", "other-test-user");
-    }
-
-    @Test
-    public void findByPrincipalAndType() {
-        persistenceAuditEventRepository.save(testUserEvent);
-        persistenceAuditEventRepository.save(testOldUserEvent);
-
-        testOtherUserEvent.setAuditEventType(testUserEvent.getAuditEventType());
-        persistenceAuditEventRepository.save(testOtherUserEvent);
-
-        PersistentAuditEvent testUserOtherTypeEvent = new PersistentAuditEvent();
-        testUserOtherTypeEvent.setPrincipal(testUserEvent.getPrincipal());
-        testUserOtherTypeEvent.setAuditEventType("test-other-type");
-        testUserOtherTypeEvent.setAuditEventDate(testUserEvent.getAuditEventDate());
-        persistenceAuditEventRepository.save(testUserOtherTypeEvent);
-
-        List<AuditEvent> events = customAuditEventRepository.find("test-user",
-            Date.from(testUserEvent.getAuditEventDate().minusSeconds(3600)), "test-type");
-        assertThat(events).hasSize(1);
-        AuditEvent event = events.get(0);
-        assertThat(event.getPrincipal()).isEqualTo(testUserEvent.getPrincipal());
-        assertThat(event.getType()).isEqualTo(testUserEvent.getAuditEventType());
-        assertThat(event.getData()).containsKey("test-key");
-        assertThat(event.getData().get("test-key").toString()).isEqualTo("test-value");
-        assertThat(event.getTimestamp()).isEqualTo(Date.from(testUserEvent.getAuditEventDate()));
-    }
-
-    @Test
     public void addAuditEvent() {
         Map<String, Object> data = new HashMap<>();
         data.put("test-key", "test-value");
@@ -167,7 +87,7 @@ public class CustomAuditEventRepositoryIntTest {
         assertThat(persistentAuditEvent.getAuditEventType()).isEqualTo(event.getType());
         assertThat(persistentAuditEvent.getData()).containsKey("test-key");
         assertThat(persistentAuditEvent.getData().get("test-key")).isEqualTo("test-value");
-        assertThat(persistentAuditEvent.getAuditEventDate()).isEqualTo(event.getTimestamp().toInstant());
+        assertThat(persistentAuditEvent.getAuditEventDate()).isEqualTo(event.getTimestamp());
     }
 
     @Test
@@ -189,7 +109,7 @@ public class CustomAuditEventRepositoryIntTest {
         String actualData = persistentAuditEvent.getData().get("test-key");
         assertThat(actualData.length()).isEqualTo(EVENT_DATA_COLUMN_MAX_LENGTH);
         assertThat(actualData).isSubstringOf(largeData);
-        assertThat(persistentAuditEvent.getAuditEventDate()).isEqualTo(event.getTimestamp().toInstant());
+        assertThat(persistentAuditEvent.getAuditEventDate()).isEqualTo(event.getTimestamp());
     }
 
     @Test
