@@ -1,18 +1,55 @@
 const webpack = require('webpack');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
 const utils = require('./utils.js');
+
+const getTsLoaderRule = env => {
+  const rules = [
+    { loader: 'cache-loader' },
+    {
+        loader: 'thread-loader',
+        options: {
+            // there should be 1 cpu for the fork-ts-checker-webpack-plugin
+            workers: require('os').cpus().length - 1
+        }
+    },
+    {
+      loader: 'ts-loader',
+      options: {
+        transpileOnly: true,
+        happyPackMode: true
+      }
+    }
+  ];
+  if (env === 'development') {
+    rules.unshift({
+      loader: 'react-hot-loader/webpack'
+    });
+  }
+  return rules;
+};
+
 module.exports = options => ({
   cache: options.env !== 'production',
   resolve: {
     extensions: [
       '.js', '.jsx', '.ts', '.tsx', '.json'
     ],
-    modules: ['node_modules']
+    modules: ['node_modules'],
+    alias: {
+      app: utils.root('src/main/webapp/app/')
+    }
   },
   module: {
     rules: [
+      {
+        test: /\.tsx?$/,
+        use: getTsLoaderRule(options.env),
+        include: [utils.root('./src/main/webapp/app')],
+        exclude: ['node_modules']
+      },
       {
         test: /\.(jpe?g|png|gif|svg|woff2?|ttf|eot)$/i,
         loaders: ['file-loader?hash=sha512&digest=hex&name=[hash].[ext]']
@@ -39,22 +76,14 @@ module.exports = options => ({
         NODE_ENV: `'${options.env}'`,
         VERSION: `'${utils.parseVersion()}'`,
         DEBUG_INFO_ENABLED: options.env === 'development',
-        // The root URL for API calls, ending with a '/' - for example: `"http://www.jhipster.tech:8081/myservice/"`.
+        // The root URL for API calls, ending with a '/' - for example: `"https://www.jhipster.tech:8081/myservice/"`.
         // If this URL is left empty (""), then it will be relative to the current context.
         // If you use an API server, in `prod` mode, you will need to enable CORS
         // (see the `jhipster.cors` common JHipster property in the `application-*.yml` configurations)
         SERVER_API_URL: `''`
       }
     }),
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      chunks: ['main'],
-      minChunks: module => utils.isExternalLib(module)
-    }),
-    new webpack.optimize.CommonsChunkPlugin({
-      name: ['manifest'],
-      minChunks: Infinity
-    }),
+    new ForkTsCheckerWebpackPlugin({ tslint: true }),
     new CopyWebpackPlugin([
       { from: './node_modules/swagger-ui/dist/css', to: 'swagger-ui/dist/css' },
       { from: './node_modules/swagger-ui/dist/lib', to: 'swagger-ui/dist/lib' },
