@@ -12,6 +12,7 @@ const ENV = 'development';
 
 module.exports = webpackMerge(commonConfig({ env: ENV }), {
   devtool: 'cheap-module-source-map', // https://reactjs.org/docs/cross-origin-errors.html
+  mode: 'development',
   entry: [
     'react-hot-loader/patch',
     './src/main/webapp/app/index'
@@ -23,20 +24,6 @@ module.exports = webpackMerge(commonConfig({ env: ENV }), {
   },
   module: {
     rules: [
-      {
-        test: /\.tsx?$/,
-        use: [{
-          loader: 'react-hot-loader/webpack',
-        },
-        {
-          loader: 'awesome-typescript-loader',
-          options: {
-            useCache: true
-          }
-        }],
-        include: [utils.root('./src/main/webapp/app')],
-        exclude: ['node_modules']
-      },
       {
         test: /\.scss$/,
         loaders: ['style-loader', 'css-loader', 'postcss-loader', 'sass-loader']
@@ -81,8 +68,6 @@ module.exports = webpackMerge(commonConfig({ env: ENV }), {
         reload: false
     }),
     new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoEmitOnErrorsPlugin(),
-    new webpack.NamedModulesPlugin(),
     new writeFilePlugin(),
     new webpack.WatchIgnorePlugin([
       utils.root('src/test'),
