@@ -2,6 +2,7 @@ const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
 const WorkboxPlugin = require('workbox-webpack-plugin');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const path = require('path');
 
 const utils = require('./utils.js');
@@ -9,10 +10,10 @@ const commonConfig = require('./webpack.common.js');
 
 const ENV = 'production';
 const extractCSS = new ExtractTextPlugin(`[name].[hash].css`);
-const extractSASS = new ExtractTextPlugin(`[name]-sass.[hash].css`);
 
 module.exports = webpackMerge(commonConfig({ env: ENV }), {
   // devtool: 'source-map', // Enable source maps. Please note that this will slow down the build
+  mode: 'production',
   entry: {
     main: './src/main/webapp/app/index'
   },
@@ -29,24 +30,6 @@ module.exports = webpackMerge(commonConfig({ env: ENV }), {
         loader: 'stripcomment-loader'
       },
       {
-        test: /\.tsx?$/,
-        use: [{
-          loader: 'awesome-typescript-loader',
-          options: {
-            useCache: false
-          }
-        }],
-        include: [utils.root('./src/main/webapp/app')],
-        exclude: ['node_modules']
-      },
-      {
-        test: /\.scss$/,
-        use: extractSASS.extract({
-          fallback: 'style-loader',
-          use: ['css-loader', 'postcss-loader', 'sass-loader']
-        })
-      },
-      {
         test: /\.css$/,
         use: extractCSS.extract({
           fallback: 'style-loader',
@@ -55,34 +38,42 @@ module.exports = webpackMerge(commonConfig({ env: ENV }), {
       }
     ]
   },
-  plugins: [
-    // this conflicts with -p option
-    new webpack.optimize.UglifyJsPlugin({
-      beautify: false,
-      comments: false,
-      // sourceMap: true, // Enable source maps. Please note that this will slow down the build
-      compress: {
-        screw_ie8: true,
-        warnings: false
-      },
-      mangle: {
-        keep_fnames: true,
-        screw_i8: true
+  optimization: {
+    runtimeChunk: false,
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          chunks: 'all'
+        }
       }
-    }),
+    },
+    minimizer: [
+      new UglifyJsPlugin({
+        cache: true,
+        parallel: true,
+        // sourceMap: true, // Enable source maps. Please note that this will slow down the build
+        uglifyOptions: {
+          beautify: false,
+          comments: false,
+          compress: {
+            warnings: false
+          },
+          mangle: {
+            keep_fnames: true
+          }
+        }
+      })
+    ]
+  },
+  plugins: [
     extractCSS,
-    extractSASS,
     new webpack.LoaderOptionsPlugin({
       minimize: true,
       debug: false
     }),
-    new WorkboxPlugin({
-      // to cache all under target/www
-      globDirectory: utils.root('target/www'),
-      // find these files and cache them
-      globPatterns: ['**/*.{html,bundle.js,css,png,svg,jpg,gif,json}'],
-      // create service worker at the target/www
-      swDest: path.resolve(utils.root('target/www'), 'sw.js'),
+    new WorkboxPlugin.GenerateSW({
       clientsClaim: true,
       skipWaiting: true,
     })
