@@ -2,10 +2,12 @@ const webpack = require('webpack');
 const path = require('path');
 const webpackMerge = require('webpack-merge');
 const commonConfig = require('./webpack.common.js');
+const utils = require('./utils.js');
 
 const ENV = 'development';
 
 module.exports = {
+  mode: 'development',
   entry: ['./src/test/javascript/spec/entry'],
   module: {
     rules: [
@@ -18,10 +20,6 @@ module.exports = {
         loaders: ['style-loader', 'css-loader']
       },
       {
-        test: /\.(jpe?g|png|gif|svg|woff2?|ttf|eot)$/i,
-        loaders: ['file-loader?hash=sha512&digest=hex&name=[hash].[ext]']
-      },
-      {
         enforce: 'pre',
         test: /\.jsx?$/,
         loader: 'source-map-loader'
@@ -29,18 +27,20 @@ module.exports = {
       {
         test: /src[/|\\]main[/|\\]webapp[/|\\].+\.tsx?$/,
         enforce: 'post',
-        exclude: /(test|node_modules)/,
+        exclude: /node_modules/,
         loader: 'sourcemap-istanbul-instrumenter-loader?force-sourcemap=true'
       },
       {
         test: /\.tsx?$/,
-        use: [{
-          loader: 'awesome-typescript-loader',
-          options: {
-            useCache: true,
-            ignoreDiagnostics: [2307] // due to a weird false error from json files
+        use: [
+          { loader: 'cache-loader' },
+          {
+            loader: 'ts-loader',
+            options: {
+              ignoreDiagnostics: [2307] // due to a weird false error from json files
+            }
           }
-        }],
+        ],
         include: [path.resolve('./src/main/webapp/app'), path.resolve('./src/test/javascript')],
         exclude: ['node_modules']
       }
@@ -51,7 +51,10 @@ module.exports = {
     extensions: [
       '.js', '.jsx', '.ts', '.tsx', '.json'
     ],
-    modules: ['node_modules']
+    modules: ['node_modules'],
+    alias: {
+      app: utils.root('src/main/webapp/app/')
+    }
   },
   plugins: [
     new webpack.SourceMapDevToolPlugin({
